+++ perfect-service/src/main/java/com/perfect/service/impl/AccountOverviewServiceImpl.java
@@ -18,58 +18,60 @@ import java.util.Map;
  * Created by john on 2014/7/25.
  */
 @Service("accountOverviewService")
-public class AccountOverviewServiceImpl implements AccountOverviewService{
+public class AccountOverviewServiceImpl implements AccountOverviewService {
 
     @Resource
     private AccountAnalyzeDAO accountAnalyzeDAO;
 
-    private  DateFormat df = new SimpleDateFormat("yyyy-MM-dd");
+    private DateFormat df = new SimpleDateFormat("yyyy-MM-dd");
+
     /**
      * 汇总。。。。
+     *
      * @return
      */
     @Override
-    public  Map<String,Object> getKeyWordSum(String startDate,String endDate){
+    public Map<String, Object> getKeyWordSum(String startDate, String endDate) {
         //各种汇总初始化
-        long impressionCount = 0 ;
-        long clickCount = 0 ;
+        long impressionCount = 0;
+        long clickCount = 0;
         double costCount = 0.0;
         double conversionCount = 0.0;
 
-         //开始获取数据汇总
+        //开始获取数据汇总
         List<AccountReportEntity> list = null;
         try {
-            list = accountAnalyzeDAO.performaneCurve(df.parse(startDate),df.parse(endDate));
+            list = accountAnalyzeDAO.performaneCurve(df.parse(startDate), df.parse(endDate));
         } catch (ParseException e) {
             e.printStackTrace();
         }
 
         for (AccountReportEntity are : list) {
-            if(are.getPcImpression()!=null){
+            if (are.getPcImpression() != null) {
                 impressionCount += are.getPcImpression();
             }
-            if(are.getMobileImpression()!=null){
+            if (are.getMobileImpression() != null) {
                 impressionCount += are.getMobileImpression();
             }
 
-            if(are.getPcClick()!=null){
+            if (are.getPcClick() != null) {
                 clickCount += are.getPcClick();
             }
-            if(are.getMobileClick()!=null){
+            if (are.getMobileClick() != null) {
                 clickCount += are.getMobileClick();
             }
 
-            if(are.getPcCost()!=null){
-                costCount += are.getPcCost();
+            if (are.getPcCost() != null) {
+                costCount += are.getPcCost().doubleValue();
             }
-            if(are.getMobileCost()!=null){
-                costCount += are.getMobileCost();
+            if (are.getMobileCost() != null) {
+                costCount += are.getMobileCost().doubleValue();
             }
 
-            if(are.getPcConversion()!=null){
+            if (are.getPcConversion() != null) {
                 conversionCount += are.getPcConversion();
             }
-            if(are.getMobileConversion()!=null){
+            if (are.getMobileConversion() != null) {
                 conversionCount += are.getMobileConversion();
             }
 
@@ -78,11 +80,11 @@ public class AccountOverviewServiceImpl implements AccountOverviewService{
         //数字格式化
         DecimalFormat decimalFormat = new DecimalFormat("#,##,###");
         DecimalFormat costFormat = new DecimalFormat("#,##,##0.00");
-        Map<String,Object> map = new Hashtable<String,Object>();
-        map.put("impression",decimalFormat.format(impressionCount));
-        map.put("click",decimalFormat.format(clickCount));
-        map.put("cos",costFormat.format(costCount));
-        map.put("conversion",decimalFormat.format((long)conversionCount));
+        Map<String, Object> map = new Hashtable<String, Object>();
+        map.put("impression", decimalFormat.format(impressionCount));
+        map.put("click", decimalFormat.format(clickCount));
+        map.put("cos", costFormat.format(costCount));
+        map.put("conversion", decimalFormat.format((long) conversionCount));
         return map;
     }
